[Fix] Updated delete user function

+++ app/controllers/userController.js
@@ -93,14 +93,32 @@ const userController = {
 
   async deletUser(req, res) {
     const userId = req.params.id;
-    const deleteUser = await userDataMapper.deleteUser(userId);
-    if (deleteUser) {
-      res.send("User successfully deleted");
+
+    // Check if user exist
+    const user = await userDataMapper.getUserById(userId);
+
+    if (!user) {
+      res.status(401).send("No User with this id in database ");
     } else {
-      res.send("Error, user can't be deleted ");
+      const oldpassword = req.body.oldpassword;
+      const fetchPassword = user.password;
+
+      const isPasswordCorrect = await bcrypt.compare(
+        oldpassword,
+        fetchPassword
+      );
+
+      if (isPasswordCorrect) {
+        const deleteUser = await userDataMapper.deleteUser(userId);
+
+        if (deleteUser) {
+          res.send("User successfully deleted");
+        } else {
+          res.send("Error, user can't be deleted ");
+        }
+      }
     }
   },
-
   async getAllInfosFromUserId(req, res) {
     const userId = req.params.id;
     const data = await userDataMapper.getAllInfosFromUserId(userId);
